+++ CHANGELOG.md
@@ -13,6 +13,9 @@ v0.21 + 1
   us to safely update a reflog with arbitrary contents, as we need to
   do for stash.
 
+* The index' tree cache is now filled upon read-tree and write-tree
+  and the cache is written to disk.
+
 * LF -> CRLF filter refuses to handle mixed-EOL files
 
 * LF -> CRLF filter now runs when * text = auto (with Git for Windows 1.9.4)